+++ npc/cli/core.py
@@ -178,7 +178,8 @@ def _make_parser():
     parser_list.add_argument('-m', '--metadata', nargs="?", const='default', default=False, help="Add metadata to the output. If the output format supports more than one metadata format, you can specify that format as well.")
     parser_list.add_argument('--title', help="Title to show in the metadata. Overrides the title from settings.", metavar="TITLE")
     parser_list.add_argument('-o', '--outfile', nargs="?", const='-', default=None, help="File where the listing will be saved")
-    parser_list.add_argument('--sort', choices=['first', 'last'], default=None, help="The sort order for characters. Defaults to the settings value 'listing.sort_by'.")
+    parser_list.add_argument('--sort', choices=['first', 'last'], dest='sort_by', default=None, help="The sort order for characters. Defaults to the settings value 'listing.sort_by'.")
+    parser_list.add_argument('--no_sort', action='store_false', dest='do_sort', help="Do not sort characters at all")
     parser_list.set_defaults(func=commands.make_list, progress=_update_progress_bar)
 
     # Dump raw character data