+++ CHANGELOG.md
@@ -17,7 +17,8 @@ Werewolves are supported! Along with the tags, character sheet, and templates fo
 * New `@werewolf` tag for specifying werewolf type and auspice on one line
 * Settings now has a `tag_defaults` key that sets default tag values for newly created characters.
 * Settings files now ignore trailing commas
-* Sorting for lists now allows you to sort by any tag or set of tags, each in ascending or descending order
+* Sort character listings by any tag or set of tags, each in ascending or descending order
+* Can turn of sorting for character listings altogether
 
 ### Changed